refactor(tui): isolate session workbench rendering
