+++ framework/tui/src/profile-shell.tsx
@@ -16,6 +16,10 @@ import {
 import { resolveListWindow } from './list-window/index.js';
 import { boundedIndex } from './navigation.js';
 import {
+  type SessionWorkbenchProps,
+  SessionWorkbenchView,
+} from './session-workbench-view.js';
+import {
   KUNGFU_CIRCULAR_STARTUP_PATTERN,
   KUNGFU_EMPTY_WORK_NEBULA_PATTERN,
   type TerminalAnimationPattern,
@@ -639,866 +643,11 @@ export function renderProfileShellSnapshot(
   return lines.slice(0, dimensions.rows).join('\n');
 }
 
-export type PlaybackTiming = {
-  eventIntervalMs: number;
-  verdictIntervalMs: number;
-};
-export type IncrementalPlayback<TEvent> = {
-  enqueue(event: TEvent): void;
-  finish(): Promise<boolean>;
-  cancel(): void;
-};
-export type WorkbenchFocus = 'session-1' | 'session-2' | 'correct' | 'failed';
-export type WorkbenchReportDetail = 'correct' | 'failed';
-export type WorkbenchActionButton<Action extends string = string> = {
-  action: Action;
-  label: string;
-};
-export type WorkbenchNextPrompt = { title: string; instruction: string };
-export type WorkbenchGuideOverlay = {
-  heading: string;
-  title: string;
-  lines: string[];
-  footer: string;
-};
-export type WorkbenchLine = {
-  session: 1 | 2;
-  source: string;
-  text: string;
-  tone: 'normal' | 'running' | 'good' | 'bad' | 'dim';
-};
-export type WorkbenchSessionBuffers = Record<1 | 2, WorkbenchLine[]>;
-export type WorkbenchScrollBack = Record<1 | 2, number>;
-export type WorkbenchCheck = {
-  id: string;
-  passed: boolean;
-  title: string;
-  meaning: string;
-};
-
-export const WORKBENCH_SESSION_BUFFER_LIMIT = 1_000;
-
-export function emptyWorkbenchSessionBuffers(): WorkbenchSessionBuffers {
-  return { 1: [], 2: [] };
-}
-
-export function appendWorkbenchSessionLines({
-  buffers,
-  scrollBack,
-  lines,
-  limit = WORKBENCH_SESSION_BUFFER_LIMIT,
-}: {
-  buffers: WorkbenchSessionBuffers;
-  scrollBack: WorkbenchScrollBack;
-  lines: WorkbenchLine[];
-  limit?: number;
-}): {
-  buffers: WorkbenchSessionBuffers;
-  scrollBack: WorkbenchScrollBack;
-} {
-  const nextBuffers: WorkbenchSessionBuffers = {
-    1: buffers[1],
-    2: buffers[2],
-  };
-  const nextScrollBack: WorkbenchScrollBack = { ...scrollBack };
-  for (const session of [1, 2] as const) {
-    const appended = lines.filter((line) => line.session === session);
-    if (appended.length === 0) continue;
-    const combined = [...buffers[session], ...appended];
-    nextBuffers[session] = combined.slice(
-      Math.max(0, combined.length - Math.max(1, limit)),
-    );
-    if (scrollBack[session] > 0) {
-      nextScrollBack[session] = Math.min(
-        nextBuffers[session].length - 1,
-        scrollBack[session] + appended.length,
-      );
-    }
-  }
-  return { buffers: nextBuffers, scrollBack: nextScrollBack };
-}
-
-export function scrollWorkbenchSession({
-  current,
-  lineCount,
-  viewportRows,
-  delta,
-}: {
-  current: number;
-  lineCount: number;
-  viewportRows: number;
-  delta: number;
-}): number {
-  return Math.max(
-    0,
-    Math.min(
-      Math.max(0, lineCount - Math.max(1, viewportRows)),
-      current + delta,
-    ),
-  );
-}
-
-export function workbenchViewportRows({
-  dimensions,
-  showHelp,
-  verdictDetail,
-}: {
-  dimensions: TerminalDimensions;
-  showHelp: boolean;
-  verdictDetail?: string;
-}): number {
-  const titleColumns = Math.max(1, Math.floor(dimensions.columns / 2) - 2);
-  const textColumns = Math.max(1, titleColumns - 2);
-  const wrappedRows = (text: string) =>
-    Math.max(1, Math.ceil(text.length / textColumns));
-  const chromeRows =
-    4 +
-    (showHelp ? 1 : 0) +
-    (verdictDetail ? 7 : 6) +
-    2 +
-    1 +
-    wrappedRows('PUBLIC ACTIVITY · SENSITIVE INTERNALS HIDDEN') +
-    wrappedRows(
-      'Mouse wheel scrolls the Session under the pointer · ↑↓ scroll focused Session',
-    );
-  return Math.max(4, terminalCanvasRows(dimensions.rows) - chromeRows);
-}
-
-export function workbenchSessionAtPoint({
-  dimensions,
-  showHelp,
-  verdictDetail,
-  column,
-  row,
-  topOffset = 0,
-}: {
-  dimensions: TerminalDimensions;
-  showHelp: boolean;
-  verdictDetail?: string;
-  column: number;
-  row: number;
-  topOffset?: number;
-}): 1 | 2 | undefined {
-  const localRow = row - topOffset;
-  const headerRows = 3 + (showHelp ? 1 : 0);
-  const verdictRows = verdictDetail ? 7 : 6;
-  const finalSessionRow = terminalCanvasRows(dimensions.rows) - verdictRows;
-  if (
-    column < 1 ||
-    column > dimensions.columns ||
-    localRow <= headerRows ||
-    localRow > finalSessionRow
-  ) {
-    return undefined;
-  }
-  return column <= Math.floor(dimensions.columns / 2) ? 1 : 2;
-}
-
-export function workbenchActionAtPoint<Action extends string>({
-  actions,
-  column,
-  row,
-  topOffset = 0,
-}: {
-  actions: readonly WorkbenchActionButton<Action>[];
-  column: number;
-  row: number;
-  topOffset?: number;
-}): Action | undefined {
-  if (row - topOffset !== 3 || column < 1) return undefined;
-  let start = 1;
-  for (const action of actions) {
-    const end = start + action.label.length + 3;
-    if (column >= start && column <= end) return action.action;
-    start = end + 2;
-  }
-  return undefined;
-}
-
-export function horizontalPointerActionAtPoint<Action extends string>({
-  actions,
-  column,
-  row,
-  targetRow,
-  startColumn = 1,
-  gap = 1,
-}: {
-  actions: readonly WorkbenchActionButton<Action>[];
-  column: number;
-  row: number;
-  targetRow: number;
-  startColumn?: number;
-  gap?: number;
-}): Action | undefined {
-  if (row !== targetRow || column < startColumn) return undefined;
-  let start = startColumn;
-  for (const action of actions) {
-    const end = start + action.label.length - 1;
-    if (column >= start && column <= end) return action.action;
-    start = end + 1 + gap;
-  }
-  return undefined;
-}
-
-export function workbenchReportAtPoint({
-  dimensions,
-  column,
-  row,
-  topOffset = 0,
-}: {
-  dimensions: TerminalDimensions;
-  column: number;
-  row: number;
-  topOffset?: number;
-}): WorkbenchReportDetail | undefined {
-  const localRow = row - topOffset;
-  const canvasRows = terminalCanvasRows(dimensions.rows);
-  if (
-    column < 1 ||
-    column > dimensions.columns ||
-    localRow < canvasRows - 3 ||
-    localRow > canvasRows - 1
-  ) {
-    return undefined;
-  }
-  return column <= Math.floor(dimensions.columns / 2) ? 'correct' : 'failed';
-}
-
-export function workbenchReportReturnAtPoint({
-  dimensions,
-  column,
-  row,
-  topOffset = 0,
-}: {
-  dimensions: TerminalDimensions;
-  column: number;
-  row: number;
-  topOffset?: number;
-}): boolean {
-  const localRow = row - topOffset;
-  const canvasRows = terminalCanvasRows(dimensions.rows);
-  return (
-    column >= 1 &&
-    column <= dimensions.columns &&
-    localRow >= canvasRows - 2 &&
-    localRow <= canvasRows
-  );
-}
-
-export function createIncrementalPlayback<TEvent>({
-  timing,
-  onEvent,
-  onAssessing,
-  isCurrent = () => true,
-  wait = (milliseconds) =>
-    new Promise<void>((resolve) => {
-      setTimeout(resolve, milliseconds);
-    }),
-}: {
-  timing: PlaybackTiming;
-  onEvent: (event: TEvent) => void;
-  onAssessing: () => void;
-  isCurrent?: () => boolean;
-  wait?: (milliseconds: number) => Promise<void>;
-}): IncrementalPlayback<TEvent> {
-  let active = true;
-  let queue = Promise.resolve();
-  return {
-    enqueue(event) {
-      queue = queue
-        .then(() => wait(timing.eventIntervalMs))
-        .then(() => {
-          if (active && isCurrent()) onEvent(event);
-        });
-    },
-    async finish() {
-      await queue;
-      if (!active || !isCurrent()) return false;
-      onAssessing();
-      await wait(timing.verdictIntervalMs);
-      return active && isCurrent();
-    },
-    cancel() {
-      active = false;
-    },
-  };
-}
-
-const WORKBENCH_FOCUS_ORDER: WorkbenchFocus[] = [
-  'session-1',
-  'session-2',
-  'correct',
-  'failed',
-];
-
-export function nextWorkbenchFocus(
-  current: WorkbenchFocus,
-  reportAvailable: boolean,
-): WorkbenchFocus {
-  const available = reportAvailable
-    ? WORKBENCH_FOCUS_ORDER
-    : WORKBENCH_FOCUS_ORDER.slice(0, 2);
-  const currentIndex = Math.max(0, available.indexOf(current));
-  return available[(currentIndex + 1) % available.length];
-}
-
-export function isWorkbenchReturnInput(input: string): boolean {
-  return (
-    input === '\r' ||
-    input === '\n' ||
-    input === '\u001b' ||
-    input === '\u007f' ||
-    input === '\b' ||
-    input === 'b' ||
-    input === 'B' ||
-    input === '\u001b[D'
-  );
-}
-
-export function sessionTitleBar({
-  session,
-  title,
-  active,
-  running,
-  columns,
-  activityFrame = 0,
-}: {
-  session: 1 | 2;
-  title: string;
-  active: boolean;
-  running: boolean;
-  columns: number;
-  activityFrame?: number;
-}): string {
-  const prefix = `${active ? '>' : ' '} S${session} · `;
-  const spinner = ['◐', '◓', '◑', '◒'][activityFrame % 4];
-  const status = running ? `${spinner} RUNNING` : 'READY';
-  const titleColumns = Math.max(0, columns - prefix.length - status.length - 1);
-  const compactTitle =
-    title.length <= titleColumns
-      ? title
-      : titleColumns > 1
-        ? `${title.slice(0, titleColumns - 1)}…`
-        : '';
-  const left = `${prefix}${compactTitle}`;
-  const gap = ' '.repeat(Math.max(1, columns - left.length - status.length));
-  return `${left}${gap}${status}`.padEnd(columns).slice(0, columns);
-}
-
-export function boundedPromptRows(
-  value: string,
-  columns: number,
-  maxRows = 2,
-): string[] {
-  const width = Math.max(1, columns);
-  const rows: string[] = [];
-  let remaining = value.trim();
-  while (remaining && rows.length < maxRows) {
-    if (remaining.length <= width) {
-      rows.push(remaining);
-      remaining = '';
-      break;
-    }
-    const space = remaining.lastIndexOf(' ', width);
-    const cut = space > 0 ? space : width;
-    rows.push(remaining.slice(0, cut).trimEnd());
-    remaining = remaining.slice(cut).trimStart();
-  }
-  if (remaining && rows.length > 0) {
-    const last = rows.length - 1;
-    rows[last] = `${rows[last].slice(0, Math.max(0, width - 1))}…`;
-  }
-  while (rows.length < maxRows) rows.push('');
-  return rows;
-}
-
-function workbenchLineColor(tone: WorkbenchLine['tone']) {
-  if (tone === 'running') return 'yellow';
-  if (tone === 'good') return 'green';
-  if (tone === 'bad') return 'red';
-  if (tone === 'dim') return 'gray';
-  return undefined;
-}
-
-function WorkbenchSessionPane({
-  session,
-  title,
-  lines,
-  active,
-  scrollBack,
-  viewportRows,
-  running,
-  titleBarColumns,
-  activityFrame,
-  footer,
-}: {
-  session: 1 | 2;
-  title: string;
-  lines: WorkbenchLine[];
-  active: boolean;
-  scrollBack: number;
-  viewportRows: number;
-  running: boolean;
-  titleBarColumns: number;
-  activityFrame: number;
-  footer: string;
-}) {
-  const sessionLines = lines;
-  const liveStart = Math.max(0, sessionLines.length - viewportRows);
-  const start = Math.max(0, liveStart - scrollBack);
-  const visible = sessionLines.slice(start, start + viewportRows);
-  return (
-    <Box
-      width="50%"
-      flexDirection="column"
-      borderStyle="round"
-      borderColor={active ? 'cyan' : 'gray'}
-      overflow="hidden"
-    >
-      <Text
-        bold
-        color={active ? 'black' : 'white'}
-        backgroundColor={active ? 'cyan' : 'gray'}
-        wrap="truncate-end"
-      >
-        {sessionTitleBar({
-          session,
-          title,
-          active,
-          running,
-          columns: titleBarColumns,
-          activityFrame,
-        })}
-      </Text>
-      <Box paddingX={1}>
-        <Text dimColor>PUBLIC ACTIVITY · SENSITIVE INTERNALS HIDDEN</Text>
-      </Box>
-      <Box
-        flexDirection="column"
-        height={viewportRows}
-        overflow="hidden"
-        paddingX={1}
-      >
-        {visible.length === 0 ? (
-          <Text dimColor>Activity will appear one event at a time.</Text>
-        ) : null}
-        {visible.map((line, index) => (
-          <Text
-            key={`${start + index}-${line.source}-${line.text}`}
-            color={workbenchLineColor(line.tone)}
-            wrap="truncate-end"
-          >
-            {String(start + index + 1).padStart(2, '0')}{' '}
-            {line.source.padEnd(11)} {line.text}
-          </Text>
-        ))}
-      </Box>
-      <Box paddingX={1}>
-        <Text dimColor>{footer}</Text>
-      </Box>
-    </Box>
-  );
-}
-
-function WorkbenchReportDetail({
-  dimensions,
-  checks,
-  detail,
-  caption,
-  interactive,
-}: {
-  dimensions: TerminalDimensions;
-  checks: WorkbenchCheck[];
-  detail: WorkbenchReportDetail;
-  caption: string;
-  interactive: boolean;
-}) {
-  const rows = checks.filter(
-    (check) => (detail === 'correct') === check.passed,
-  );
-  const correct = detail === 'correct';
-  return (
-    <Box
-      width={dimensions.columns}
-      height={terminalCanvasRows(dimensions.rows)}
-      flexDirection="column"
-      borderStyle="double"
-      borderColor={correct ? 'green' : 'red'}
-      paddingX={1}
-      overflow="hidden"
-    >
-      <Text bold color={correct ? 'green' : 'red'}>
-        {interactive
-          ? `${correct ? '✓ CORRECT CHECKS' : '× FAILED CHECKS'} · ${rows.length}`
-          : correct
-            ? `✓ ACCEPTANCE REPORT · ${rows.length}/${checks.length} CHECKS PASSED`
-            : `× ACCEPTANCE REPORT · ${rows.length} FAILED CHECKS`}
-      </Text>
-      <Text dimColor>{caption}</Text>
-      <Box flexDirection="column" flexGrow={1} minHeight={0} marginTop={1}>
-        {rows.length === 0 ? (
-          <Text color={correct ? 'yellow' : 'green'}>
-            {correct
-              ? 'No correct checks were recorded.'
-              : 'No failed checks. This is the expected result.'}
-          </Text>
-        ) : null}
-        {rows.map((row, index) => (
-          <Box key={row.id} flexDirection="column" marginBottom={1}>
-            <Text bold color={row.passed ? 'green' : 'red'}>
-              {String(index + 1).padStart(2, '0')} {row.passed ? '✓' : '×'}{' '}
-              {row.title}
-            </Text>
-            <Text dimColor>{row.meaning}</Text>
-          </Box>
-        ))}
-      </Box>
-      <Box borderStyle="round" borderColor="cyan" paddingX={1}>
-        <Text bold color="cyan" wrap="truncate-end">
-          {interactive
-            ? '← RETURN TO RESULT CARDS · Esc / Enter / Backspace / b'
-            : 'DEMO COMPLETE · This acceptance report closes automatically'}
-        </Text>
-      </Box>
-    </Box>
-  );
-}
-
-function WorkbenchResultCard({
-  kind,
-  count,
-  active,
-  available,
-  emphasized,
-  interactive,
-}: {
-  kind: WorkbenchReportDetail;
-  count: number;
-  active: boolean;
-  available: boolean;
-  emphasized: boolean;
-  interactive: boolean;
-}) {
-  const correct = kind === 'correct';
-  const tone = correct || count === 0 ? 'green' : 'red';
-  const cardColor = !available ? 'gray' : active || emphasized ? 'cyan' : tone;
-  return (
-    <Box
-      width="50%"
-      height={3}
-      borderStyle={emphasized ? 'double' : 'round'}
-      borderColor={cardColor}
-      paddingX={1}
-      overflow="hidden"
-    >
-      <Text bold color={cardColor} wrap="truncate-end">
-        {active ? '> ' : '  '}
-        {correct ? '✓' : '×'} {count} {correct ? 'CORRECT' : 'FAILED'}
-        {available
-          ? interactive
-            ? ' · click / Enter details'
-            : ' · verified'
-          : ' · waiting'}
-      </Text>
-    </Box>
-  );
-}
-
-function opaqueWorkbenchLine(value: string, columns: number): string {
-  return value.slice(0, columns).padEnd(columns);
-}
-
-export type SessionWorkbenchProps = {
-  dimensions: TerminalDimensions;
-  heading: string;
-  collectionLabel: string;
-  caseLabel: string;
-  relationship: string;
-  controls: string;
-  controlActions?: WorkbenchActionButton[];
-  help: string;
-  sourceLabel: string;
-  targetLabel: string;
-  buffers: WorkbenchSessionBuffers;
-  checks: WorkbenchCheck[];
-  reportAvailable: boolean;
-  reportPassed: boolean;
-  verdictSuccess: string;
-  verdictFailure: string;
-  verdictDetail?: string;
-  detailCaption: string;
-  busy: string;
-  progress: string;
-  error: string;
-  activeFocus: WorkbenchFocus;
-  scrollBack: WorkbenchScrollBack;
-  showHelp: boolean;
-  activityFrame: number;
-  runningSession?: 1 | 2;
-  nextPrompt?: WorkbenchNextPrompt;
-  guideOverlay?: WorkbenchGuideOverlay;
-  reportDetail?: WorkbenchReportDetail;
-  emphasizedResult?: WorkbenchReportDetail;
-  interactive?: boolean;
-};
+export * from './session-workbench-state.js';
+export type { SessionWorkbenchProps } from './session-workbench-view.js';
 
 export function SessionWorkbench(props: SessionWorkbenchProps) {
-  const {
-    dimensions,
-    heading,
-    collectionLabel,
-    caseLabel,
-    relationship,
-    controls,
-    controlActions,
-    help,
-    sourceLabel,
-    targetLabel,
-    buffers,
-    checks,
-    reportAvailable,
-    reportPassed,
-    verdictSuccess,
-    verdictFailure,
-    verdictDetail,
-    detailCaption,
-    busy,
-    progress,
-    error,
-    activeFocus,
-    scrollBack,
-    showHelp,
-    activityFrame,
-    runningSession,
-    nextPrompt,
-    guideOverlay,
-    reportDetail,
-    emphasizedResult,
-    interactive = true,
-  } = props;
-  if (reportAvailable && reportDetail) {
-    return (
-      <WorkbenchReportDetail
-        dimensions={dimensions}
-        checks={checks}
-        detail={reportDetail}
-        caption={detailCaption}
-        interactive={interactive}
-      />
-    );
-  }
-  const titleColumns = Math.max(1, Math.floor(dimensions.columns / 2) - 2);
-  const viewportRows = workbenchViewportRows({
-    dimensions,
-    showHelp,
-    verdictDetail,
-  });
-  const passedCount = checks.filter((check) => check.passed).length;
-  const failedCount = checks.length - passedCount;
-  const promptWidth = Math.min(
-    dimensions.columns,
-    Math.min(68, Math.max(24, dimensions.columns - 8)),
-  );
-  const promptColumns = Math.max(1, promptWidth - 2);
-  const promptRows = nextPrompt
-    ? boundedPromptRows(
-        `${nextPrompt.title} · ${nextPrompt.instruction}`,
-        Math.max(1, promptColumns - 2),
-      )
-    : [];
-  const guideWidth = Math.min(
-    dimensions.columns,
-    Math.min(88, Math.max(32, dimensions.columns - 8)),
-  );
-  const guideColumns = Math.max(1, guideWidth - 2);
-  const guideRows =
-    guideOverlay?.lines.flatMap((line) =>
-      boundedPromptRows(line, Math.max(1, guideColumns - 2)),
-    ) ?? [];
-  const sessionFooter = (session: 1 | 2) =>
-    interactive
-      ? `click focus · wheel here / ↑↓ scroll · Tab switch · ${
-          scrollBack[session] > 0
-            ? `${scrollBack[session]} lines back`
-            : 'following live'
-        }`
-      : 'Following admitted public activity one event at a time.';
-  return (
-    <Box
-      width={dimensions.columns}
-      height={terminalCanvasRows(dimensions.rows)}
-      flexDirection="column"
-      overflow="hidden"
-    >
-      <Box paddingX={1} justifyContent="space-between">
-        <Text bold color="cyan">
-          {heading.toUpperCase()}
-        </Text>
-        <Text>
-          {collectionLabel} · {caseLabel}
-        </Text>
-      </Box>
-      <Text wrap="truncate-end">
-        S1 {sourceLabel} {relationship} S2 {targetLabel}
-      </Text>
-      <Text dimColor wrap="truncate-end">
-        {controlActions?.map((action, index) => (
-          <React.Fragment key={action.action}>
-            {index > 0 ? ' ' : null}
-            <Text bold color="cyan">
-              [ {action.label} ]
-            </Text>
-          </React.Fragment>
-        ))}
-        {controlActions?.length ? ' · ' : null}
-        {controls}
-      </Text>
-      {showHelp ? (
-        <Text dimColor wrap="truncate-end">
-          {help}
-        </Text>
-      ) : null}
-      <Box flexGrow={1} minHeight={0}>
-        <WorkbenchSessionPane
-          session={1}
-          title={sourceLabel}
-          lines={buffers[1]}
-          active={activeFocus === 'session-1'}
-          scrollBack={scrollBack[1]}
-          viewportRows={viewportRows}
-          running={Boolean(progress) && runningSession === 1}
-          titleBarColumns={titleColumns}
-          activityFrame={activityFrame}
-          footer={sessionFooter(1)}
-        />
-        <WorkbenchSessionPane
-          session={2}
-          title={targetLabel}
-          lines={buffers[2]}
-          active={activeFocus === 'session-2'}
-          scrollBack={scrollBack[2]}
-          viewportRows={viewportRows}
-          running={Boolean(progress) && runningSession === 2}
-          titleBarColumns={titleColumns}
-          activityFrame={activityFrame}
-          footer={sessionFooter(2)}
-        />
-      </Box>
-      <Box
-        height={verdictDetail ? 7 : 6}
-        borderStyle="round"
-        borderColor={
-          reportAvailable ? (reportPassed ? 'green' : 'red') : 'gray'
-        }
-        paddingX={1}
-        flexDirection="column"
-        overflow="hidden"
-      >
-        <Text
-          color={reportAvailable ? (reportPassed ? 'green' : 'red') : 'yellow'}
-          bold
-          wrap="truncate-end"
-        >
-          {reportAvailable
-            ? reportPassed
-              ? verdictSuccess
-              : verdictFailure
-            : error || busy || progress || 'Ready · choose a test case'}
-        </Text>
-        {reportAvailable && verdictDetail ? (
-          <Text color={reportPassed ? 'green' : 'red'} bold wrap="truncate-end">
-            {verdictDetail}
-          </Text>
-        ) : null}
-        <Box>
-          <WorkbenchResultCard
-            kind="correct"
-            count={passedCount}
-            active={interactive && activeFocus === 'correct'}
-            available={reportAvailable}
-            emphasized={emphasizedResult === 'correct'}
-            interactive={interactive}
-          />
-          <WorkbenchResultCard
-            kind="failed"
-            count={failedCount}
-            active={interactive && activeFocus === 'failed'}
-            available={reportAvailable}
-            emphasized={emphasizedResult === 'failed'}
-            interactive={interactive}
-          />
-        </Box>
-      </Box>
-      {nextPrompt ? (
-        <Box
-          position="absolute"
-          width={promptWidth}
-          height={6}
-          marginTop={Math.max(4, Math.floor(dimensions.rows / 2) - 2)}
-          marginLeft={Math.max(
-            2,
-            Math.floor((dimensions.columns - promptWidth) / 2),
-          )}
-          borderStyle="double"
-          borderColor="yellow"
-          flexDirection="column"
-          overflow="hidden"
-        >
-          <Text bold color="yellow" backgroundColor="blue">
-            {opaqueWorkbenchLine(' WHAT TO TRY NEXT', promptColumns)}
-          </Text>
-          {promptRows.map((row, index) => (
-            <Text key={`${index}-${row}`} color="white" backgroundColor="blue">
-              {opaqueWorkbenchLine(` ${row}`, promptColumns)}
-            </Text>
-          ))}
-          <Text color="white" backgroundColor="blue">
-            {opaqueWorkbenchLine(
-              ' Closes automatically in 5 seconds.',
-              promptColumns,
-            )}
-          </Text>
-        </Box>
-      ) : null}
-      {guideOverlay ? (
-        <Box
-          position="absolute"
-          width={guideWidth}
-          height={Math.min(
-            terminalCanvasRows(dimensions.rows) - 4,
-            guideRows.length + 5,
-          )}
-          marginTop={Math.max(
-            2,
-            Math.floor((dimensions.rows - guideRows.length - 5) / 2),
-          )}
-          marginLeft={Math.max(
-            2,
-            Math.floor((dimensions.columns - guideWidth) / 2),
-          )}
-          borderStyle="double"
-          borderColor="cyan"
-          flexDirection="column"
-          overflow="hidden"
-        >
-          <Text bold color="black" backgroundColor="cyan">
-            {opaqueWorkbenchLine(` ${guideOverlay.heading}`, guideColumns)}
-          </Text>
-          <Text bold color="cyan" backgroundColor="black">
-            {opaqueWorkbenchLine(` ${guideOverlay.title}`, guideColumns)}
-          </Text>
-          {guideRows.map((row, index) => (
-            <Text key={`${index}-${row}`} color="white" backgroundColor="black">
-              {opaqueWorkbenchLine(` ${row}`, guideColumns)}
-            </Text>
-          ))}
-          <Text bold color="cyan" backgroundColor="black">
-            {opaqueWorkbenchLine(` ${guideOverlay.footer}`, guideColumns)}
-          </Text>
-        </Box>
-      ) : null}
-    </Box>
-  );
+  return <SessionWorkbenchView {...props} />;
 }
 
 function controlPlaneKindColor(kind: ProductSearchDocument['kind']): string {
